Force wakeup to be deasserted on all cores after pm start
The last operation we do is a full reset of the pm complex.
That leaves wakeup enabled and ATTR_SPCWKUP_COUNT=1 on all of
the cores.  When we get to runtime we will also first do a
resetpmcomplex before we load everything.  That would then
put the ATTR_SPCWKUP_COUNT=2.  So after that we're off by one
and will never clear it on the ECO cores.  The non-ECO cores
are not affected by this issue because we use PHYP to handle
the counters there and they will start at zero when we get to
runtime.

Change-Id: Iba2ce1c4f580627895be750da630ec31d731aaf2
CQ: SW543301
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127237
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>