Added support for Version 1 of the trace entry binary data
Currently version 2 of the trace entry is being read by this file.
Adding back support for version 1 of the trace entry.  When a version 1
of binary file or binary data is given, this script will read out
the traces and put the traces into a file, or onto the console depending
on the input parameters when script used as an executable.  A string is
passed back when using API get_binary_trace_data_as_string when
script is used as a library.

Change-Id: I02b37b543003ab051d28e3ce1ee1b74dc3198b9b
RTC:295909
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127357
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>