Remove swap of "-" to "." in HDAT MTM Handling
This commit removes a previous HDAT change where the Machine Type
Model read from VPD was having its "-" replaced with a ".".  This
changed value was then stored in an attribute used by the error log
code.  This change should keep the MTM value the same throughout the
full IPL of the system.

Change-Id: Ib0274704fab7a3b5efca8a44161cd74ea9b0a605
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127420
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>