Use FRU Part Number for FRU Callouts
There are two similarly named VPD fields to describe a FRU.  There
is Part Number (PN) and FRU Part Number (FN).  Per the PEL spec,
FRU callouts should be using FN.

As part of this change there is also a refactor of the interfaces
used to determine the vpd record+keyword combinations for PN, SN,
CCIN, and FN.  It is now more extensible versus adding a new
interface or new argument each time.

Change-Id: I2d56e2d6955881e3f4feccb809758775465c751c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127481
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>