Notify PHYP before resetting PM Complex
There is a potential conflict if PHYP is modifying the HOMER
memory while it is being reloaded.  To prevent this we will
inform PHYP before we put the PM Complex into reset, as this
always happens before any sort of reload.  They will then stop
any HOMER modifications until the complex is up and alive again
(signalled via OCC interrupt).

The interface is via a new firmware_request type of
HBRT_FW_MSG_TYPE_PMRESET_ALERT=20 that takes a single argument
for the processor chip.

Change-Id: I931022b4fc17a4bc1022c0eb053fc4ae6907d095
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127433
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>