Make all unsupported SBE PSU command errors visible
BACKGROUND

SBE firmware may not support all the SBE PSU commands that Hostboot
sends it. If it doesn't support a command, it will reply with a
special return code indicating that the command is not supported.

For some commands, firmware can continue the IPL even when the command
is not supported, and for others, it has to halt the IPL. Up to now,
Hostboot has created non-visible error logs for certain unsupported
SBE PSU operations when it can continue to boot without the SBE
support. This is not desirable because all the SBE PSU operations
implemented by Hostboot so far should be supported by contemporary SBE
firmwares, and an "unsupported command" error at this stage means that
something is wrong. We don't want these problems to go unnoticed.

However, since an SBE update could fix a problem caused by outdated
SBE firmware, we don't want to log errors for all unsupported commands
until we are sure that we are running the latest SBE firmware.

THIS COMMIT

This commit changes the performPsuChipOp interface to create visible
error logs when callers specify that command support is optional. If
not returned to the caller, these errors will be added to a list of
errors that will all be committed if we pass the SBE update IStep
without updating an SBE and rebooting.

Callers can still request that "unsupported command" errors be
non-visible. This will be necessary when Hostboot implements calls to
new SBE PSU operations that may not be implemented by the SBE until
later. This will avoid creating prerequisite dependencies from
Hostboot commits to SBE commits until necessary. (However, all
currently implemented SBE PSU ops create visible errors because they
should have been implemented by modern SBE firmwares by now.)

RTC: 270168
Change-Id: I7f3be8bee2a10b6b6fddad98559e9ddef1c76298
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127472
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>