Add external patch support to bmc patch tool
A common scenario is that a developer will generate a set of
patches that need to applied on multiple systems at varying
times.  These changes will allow a developer to copy the
appropriate files to an external directory (e.g. gsa) and
then have a non-developer apply those patches to the BMC.

Another tweak was made to copy the symbol files needed for
dump debug and patch files to identify the changes that
are being applied.

Change-Id: I697f0d34057946e9a4c03e6d1613470bd38aa0e0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127484
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>