Handle inbound PLDM requests in HBRT during request handling
If BMC sends a new PLDM request while HBRT is waiting for a response to
its own PLDM request, HBRT currently queues the inbound request but
does not process it.  If this happens multiple times in a row, HBRT will
start reporting errors that the inbound request queue is full.  Worse,
since HBRT already put an inbound message on the queue, nothing will wake
HBRT up to process it.  This commit forces a discard of any queued request
message if BMC sends another one during a request loop, on the premise that
it can only do so if it is timing out a previous request (the one we're
holding on the queue).  Additionally, if HBRT receives the response to a
message it previously sent, and in the course of doing so, finds that
a request message from BMC is pending, it will ask PHyp to call HBRT back
with an MCTP message available message, which will trigger HBRT to process
the pending message.  The overall ramification from a BMC perspective is that
PLDM messages destined to HBRT can time out at any time since a) HBRT might
not even be running b) HBRT might be invoking a long running operation like
starting the PM complex and can't service inbound requests or c) HBRT might
be stuck in an up to 90s loop waiting for a response to a prior request and
cannot respond to a new inbound request at that time.  BMC should interpret
an HBRT message timeout as HBRT being transiently busy, and retry the
operation (perhaps with exponential backoff) until the operation succeeds.

Change-Id: I4a75acc9d518bf9169c93efc401821cbd3000c1c
CQ: SW543707
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127485
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>