p10_fbc_eff_config_links -- skip aggregate check in SMP wrap mode
This commit fixes a SMP wrap mode specific regression introduced
in 127007.  In that commit, the determination for the presence
of aggregate links was shifted into p10_fbc_eff_config_links,
from p10_fbc_eff_config_aggregate -- the code refactor neglected
to consider that parallel links are present and trained in wrap
mode, but are not intended to be configured in an aggregate
fashion.

This commit guards the execution of the aggregate determination
code by a check for the presence of wrap mode, which exists
already in the original p10_fbc_eff_config_aggregate code

Change-Id: Ib784e23b7d8dcb31f84712df2ad2ab4bf5ff6e4f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127544
Dev-Ready: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127548
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>