Dont skip NODE presence detection in platPresenceDetect
CONFIG_HAVE_PVPD used to be the key on which to do node presence detect
for systems. However, that flag is no longer meaningful and is unset in
P10. This leads to missing CCIN, PN, and SN attributes on the node which are
required to be filled for eBMC instances of Hostboot since Hostboot does
the FRU callouts for eBMC.

This commit removes the special case logic for the node for that config
flag in platPresenceDetect and switches to relying on hasRemoteVpdSource to
provide the correct answers in key places to fill those ATTRs for the
node. The calls to hasRemoteVpdSource were already in place so only the
removal of the config flag was necessary to get those ATTRs populated
without proceeding down paths the flag intended to prevent the node from
going down in the first place.

Change-Id: Ie19461eb3d6284f9a22fe5dc48dae1803c8c2b0c
CQ:SW543283
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127636
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>