Call PRD at the end of istep 9.1
PLL errors are injected before istep 9, causing downstream attention on
the fabric buses. PRD is first called at the end of 9.6. Unfortunately,
the HWPs in the middle of step 9 will see the downstream attentions,
callout the processors, and terminate the IPL before PRD can handle the
PLL attention. Adding PRD analysis at the end of 9.1 to potentially
catch the PLL attentions. Avoiding moving PRD into the middle of the
fabric training isteps (9.2-9.5) because the HWPs may leave lingering
attentions that are eventually cleaned up in 9.6.

Change-Id: I7f1139722087e4d47b27f2e23b4e621ec82b039d
CQ: SW543091
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127679
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>