Set flag in TI area to indicate visibility of BMC log
Hostboot can TI with either an error log ID or a pure SRC that doesn't
correspond to an errorlog. Certain SRCs need the BMC to dump Hostboot
state, and some need the BMC to produce a visible errorlog to record
the action.

This commit adds a new flag to the TI area to indicate whether a
visible error log should be created by the BMC. It also adds a table
describing which SRCs should ask for Hostboot dumps, and which should
ask for visible error logs. The BMC reads the TI area when Hostboot
TIs and will look at the flags to take action.

Change-Id: I4a3e40c8a20af7cfb5f48d27622edc849c262cf2
RTC: 297849
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127808
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>