Fixed issue with backtraces for component ID 0x0100
The file b0100.py, that deciphers the backtraces, was having
and issue locating the hbicore.syms file, known as LID 81e00686,
and unable to decipher the backtraces for display.

Added miscUtil.py which contains API 'getLid' that when given a
LID file name will locate the file, on the file system, using pre-
defined search directories.

The API in the b0100.py has been updated to use the 'getLid' API
to locate the file.

Change-Id: I8ea82fa7ca15440c9b9d5f85a73e31927237a143
RTC: 295909
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127857
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>