Reset ATTR_RECONFIGURE_LOOP before checkMinimumHardware() in host_gard
If targets with a PLID are deconfigured as a result of host_gard,
this triggers ATTR_RECONFIGURE_LOOP to be set, which will cause a
reconfig loop at the end of the istep. This is undesirable in
host_gard if checkMinimumHardware() fails, because collectGard()
will attempt to Resource Recover targets if necessary. If
collectGard() is unable to Resource Recover enough targets to
pass checkMinimumHardware(), then the system cannot ipl and will
be caught in an infinite reconfig loop.

This commit moves resetting ATTR_RECONFIGURE_LOOP before
checkMinimumHardware() in host_gard to avoid the infinite reconfig
loop scenario above and instead terminate the ipl.

Change-Id: I86a322362cdc83566a4bf522efd8a8fb2a70e9f7
CQ:SW544519
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127961
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>