Force reconfig loop when clock deconfig is requested
In general any deconfiguration during the IPL will end the istep
and trigger a reconfig loop (reboot).  However, this only works
if there is an actual target to deconfigure.  In the case of
clocks there is not a target.  Therefore if we detect a request
to deconfigure a clock we will artificially set the flag to force
a reconfig loop.

Special note - There is a possibility of an infinite loop with
this change because nothing will actually be deconfigured.  We
are going to live with this scenario because any real clock
error should be rediscovered when clock_test runs on the reboot.
At that point the appropriate deconfigs/guards/failovers/etc
will take place.

Change-Id: I7d5bc51fa75d156823a89eb22309533e2feb9cc6
CQ: SW544727
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128075
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>