Support Graceful Reboot on BMC systems with FCO (1/3)
Currently, when Field Core Override is set on BMC, HB
will restrict the number of cores to run by setting
their HWAS_STATE functional to 0 (ie not functional).
This creates a problem when FCO is set to a low number
of cores and then a Graceful reconfig loop is triggered.
BMC preserves HB's HWAS_STATE during Graceful reboots
and hands it off to SBE. This leads to undesirable
behavior since the cores knocked out by FCO will not be
considered by SBE to boot with even though they are
still technically functional.

To remedy this, HB will add anther bit in HWAS_STATE,
functionalOverride, which will be set for cores with
deconfigured by FCO in addition to setting HWAS_STATE
functional = 0.
ie:
    core0HwasState.deconfiguredByEid
                   = DECONFIGURED_BY_FIELD_CORE_OVERRIDE
    core0HwasState.functional = 0
    core0HwasState.functionalOverride = 1

BMC will key off of functionalOverride to reset target's
HWAS_STATE functional = 1. This is will make FCO cores
available for SBE to boot with. HB will do the same in
istep 6 to stay in sync with SBE.

This commit adds the new 'functionalOverride' bit to
HWAS_STATE

Change-Id: I0644eb57df6df6f9d7b9fa2d99281f338abe5ba9
CQ:SW542272
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128135
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>