checkCriticalResources() now also checks specdeconfig
This commit passes the PredicateHwas used in checkMinimumHardware()
when calling checkCriticalResources(). Before checkCriticalResources()
was using its own predicate check which did not account for specdeconfig.
This is not optimal as this would incorrectly pass the min HW check when
speculatively applying recoverable guard records. Since successful, the
recoverable records would then be applied for real and fail a later min
HW check in host_gard to fail, forcing a reboot on BMC systems to
preform Resource Recovery. With this change, the min HW fail will be
caught earlier so that HB can perform Resource Recovery without having
to reboot.

Change-Id: I3580f02ef5af71385ae7eac96b983517f8fe6c7f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128192
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>