Perform MSL checks for chip EC levels
Adds support for checking the EC levels of all chips in the system
against the MRW defined MSL levels. The list of allowed MSL EC levels is
determined by either ATTR_MSL_FIELD_SUPPORTED or ATTR_MSL_MFG_ALLOW.
Which to use is controlled by the MFG_FLAGS_MNFG_MSL_CHECK flag.

These checks are performed in istep 6.6 just before the processor EC
levels are checked for any mismatches (which this new logic doesn't
handle). This allows the system to deconfig any offending chips before
that logic and before minimum hardware check.

This new support is only for eBMC based systems as there is existing
logic in HWSV which handles the FSP based system case. However, it was
placed in common code and could potentially be commonized to both system
types in the future.

Change-Id: I96fb7a8e380a7d903495dcc49f3e96b31ad3f966
RTC:251152
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128422
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>