Expand memory target numbering to account for Odyssey
Explorer has a single MEM_PORT per OCMB, but Odyssey has 2.
These changes will expand the HUID and FAPI numbering to account
for the space Odyssey will eventually require.  Note that neither
the PHYS_PATH or AFFINITY_PATH are affected by this change as
all numbers there are relative to the parent.

I also corrected a previous mistake where we didn't allocate
space for a double-drop dimm behind the OCMB.  With this change
and the one above, the dimm FAPI_POS is 4x the parent OCMB.

The standalone simics xml was modified to reflect the numbers
we now generate from the MRW processing.

A few minor improvements were made along the way to improve
debug of future similar changes.

Change-Id: I7578db960a51144cce8e7d69ada3524592f92141
RTC: 299070
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128303
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>