Fix TPM Attestation Issue with HBB and HBBL
The Hostboot Bootloader (HBBL) extends a hash of the Hostboot Base Image
(HBB) during its part of the IPL.  HBB then needs to log this attestation
in the TPM log during its part of the IPL (because the HBBL is unable to do
this itself). There is a difference to how things get extended/logged by the
HBBL based on whether or not security is enabled.

This commit does the following things:

- In non-secure mode the HBBL will now measure the HBB image that it will
be copying into the cache and starting.  The hash of this measurement will
be copied into the secure header's hash section and extended into the TPM.

- In non-secure mode the HBB code will use the hash in the secure header
to update the TPM log.  (This was already being done in secure mode.)

- Both in secure and non-secure mode, the HBBL will only copy the
protected size of the HBB as designated by the secure header.

Change-Id: I2d495ae1a26136a8b6f927741527ad6d915086a8
CQ:SW543119
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128368
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chris Engel <cjengel@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>