Fix eBMC standalone setup script early terminations
There was a bug in the setup script where calls to usage weren't exiting
the script which would cause it to continue along until it encountered
an error or otherwise completed. This leads to annoying half-baked
branch/repo states that have to be cleaned up by the user.

This issue happened because the main script body is encapsulated in a
function so when usage calls return it goes back to 'main'. The 'exit'
keyword can't be used because this script executes in the users current
shell which would cause that shell to exit on error. So, to fix the
issue an additional 'return' is added in the places where the script was
originally supposed to quit.

Change-Id: Ie48324cc537e5cd6cae194838ff2fa31ff98f175
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128326
Reviewed-by: Chen Du <duchen@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>