HTMGT: stop processing when OCC needs a reset
There are a couple places in the code where HTMGT was not checking if
the OCC had requested a reset. This would cause longer delays before the
OCC recovery process would get started (especially when the OCCs were
first started).

If HTMGT times out waiting for the OCC to reach its final checkpoint,
HTMGT will still try sending a POLL command to check if the OCC is
functional before attempting recovery.

Additional Cleanup:
- remove setMode calls (mode is owned by BMC)
- mark OCC as failed if it does not reach checkpoint
- allow user power cap down to soft minimum limit
- clean up old TODOs
- updated the HTMGT pel parser for BMC

Change-Id: If788dfd7a7be815d9f76ba35af3b3448058fe0f3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128417
Reviewed-by: Martha Broyles <mbroyles@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sheldon Bailey <baileysh@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>