HDAT: Dead core support
This commit contains the HDAT support to handle the dead cores. If
a dead core is detected, the staus flag of that entry is explicitly
marked as Unuseable and pass the data to hypervisor instead of
completely filtering out from the core list.

Testing of this change was done by executing a MPIPL flow and observed
that dead core was marked as Unuseable and system reached back to
runtime.

Change-Id: Ie3a68729bc5e3ce9f6be4e26aa20e66e80d78fdb
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128477
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: Sampa Misra <sampmisr@in.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>