Fix an ill-formed print string that causes an error when parsing
The print string is missing the format specifier.  It has the percent
symbol (%) but no specifier on what type the argument to be printed.
This is being caught by the hostfw_trace.py tool when it tries to parse
the print string and stumbles when it hits these "naked" percent symbols.

Change-Id: Ied0666f275ebd675bb2d1edc9945c058fe904626
CQ:SW545423
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128480
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>