hostfw_trace.py: Minor bug fix and more feedback to the callee
There was a bug discovered when a char format specifier was encountered
after a long format specifier.  A flag, longflag, was being set when the long
format specifier was encountered but not being reset.  This would
cause hostfw_trace.py to choke when it encountered the char format
specifier and produce garbage.  This has been fixed by resetting the
flag at the right location.

When there was an issue with a hash string in the string file
an error was spewed to the console:
"Error: unsupported format specifier in trace found:  /"
Which was causing confusion.  Now this is a warning within the trace itself and
the ill-formed hash is spewed out so it can be corrected:
"Warnings": [
        "Warning: unsupported format specifier in trace found '/' :",
        "I>PRD Signature %/08llX %08llX"
    ],

If an empty buffer was encountered, the header string was printed out with no
trace info.  It can't be determine if it was an empty buffer or an issue:
"Data": [
"-----------------------------------------------------------------------------",
"TRACEBUFFER: tracBINARY",
"-----------------------------------------------------------------------------",
"   Sec      Usec    PID  Comp          Line Entry   Data",
"-----------------------------------------------------------------------------",
""

Now there is feedback that states it is an empty buffer along with the original
binary data so it can be examined to be confirmed it is indeed an empty buffer
or an error with the decoding that warrants further investigation:
"Issue with Data": "Buffer is empty. Data has no parseable traces",
"Data": [
    "01 28 00 42 46 41 50 49  5F 44 42 47 00 00 00 00  |  .(.BFAPI_DBG....",
    "00 00 00 00 00 00 00 28  00 00 00 00 00 00 00 28  |  .......(.......(",
    "00 00 00 00 00 00 00 00                           |  ........"
]

Give feed back when data is unparseable because the string file cannot
be found and spew back binary data:
    "File not found": "81e00685.lid",
    "Issue with Data": "Data is unparseable without file",
    "Data": [
      "01 28 00 42 50 52 44 46  00 00 00 00 00 00 00 00  |  .(.BPRDF........",
       ....

Give feed back when data is unparseable because of other reasons and
spew back binary data:
    "Decode file used": "/usr/local/share/hostfw/running/81e00685.lid",
    "Errors": [
        "Error: Unknown string file version 3",
        "Error processing string file: /usr/local/share/hostfw/
                                                          running/81e00685.lid"
    ],
    "Issue with Data": "Data is unparseable with errors",
    "Data": [
      "01 28 00 42 50 52 44 46  00 00 00 00 00 00 00 00  |  .(.BPRDF........",
       ....

Change-Id: I8e1ae283f5d7cb9dbeb15b7e05c068259c3a9ee5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128794
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>