Fix unique_ptr<T>::operator[] for GCC 9 and 10
This commit makes a small change to fix operator[] for the unique_ptr
class for GCC 9 and 10. On GCC 9 and 10, operator[] was not returning
a reference and our testcases were failing to compile, even though
GCC 8 does compile it to return a reference.

Also adds a testcase for const unique_ptr and unique_ptr-to-const.

Resolves #204
Change-Id: Ia8bca57fd70656446eb81f8bcca4122cb4d280ff
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128882
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>