PLDM Runtime BMC R/R support: outgoing PLDM request retries (commit 1/n)
This commit adds PLDM retry support to the runtime
outgoing-request-sending implementation. If an outgoing request times
out or otherwise errors, the code will retry sending the request once
before logging the error. This is so that if we make a request to the
BMC and it performs a reset/reload (R/R) and can't respond to a
request, we will retry that request to give it another chance before
giving up.

This commit does NOT add support for handling incoming request retries.

Change-Id: I81b92473864d3fb30cbcb740917ec5d7509788b3
RTC: 299918
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128956
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>