Change detection for parts with BMC-sourced VPD
There are FRUs where Hostboot does not have physical access to
the VPD EEPROMs.  For these parts we get the data from the BMC
via PLDM.  To detect a replaced part we compare the data we get
from the BMC with the data we have saved to our EECACHE.  However,
we have recently learned that neither the order of the records and
keywords we get via PLDM nor the extent of the contents (e.g. the
subset of records that are sent) are guaranteed from boot to boot.

To address this issue we will no longer rely exclusively on the
buffer contents to match exactly.  Instead we will directly compare
the SN/PN of the data in EECACHE with the data from the BMC.  To
do this we will:
1) Save the EECACHE version of the SN/PN into attributes.
2) Update EECACHE with the data from PLDM.
3) Compare saved attributes with SN/PN from new EECACHE contents.

Code Flows
cacheRemoteFruVpd()
> EEPROM::cacheEepromBuffer(target,present=true,ipzdata)
>> deviceRead(target,ipzdata,CACHE_EEPROM_ADDRESS(pres=true,VPD_AUTO))
>>> genericEepromCache(READ,target,ipzdata,EEPROM_CACHE,pres=true,VPD_AUTO)
>>>> cacheEeprom(target,pres=true,VPD_AUTO,ipzdata,size,normal_update=true)
>>>>> findEepromHeaderInPnorEecache(target,present=true,VPD_AUTO,
brand new part / empty eecache
->>>>> createNewEecacheEntry
 >>>>>> updateEecacheContents
 >>>>>>> HWAS::markTargetChanged
or for existing part
->>>>> updateExistingEecacheEntry
 >>>>>> checkForEecacheEntryUpdate
 >>>>>>> memcmp(eecache,ipz_record)
new code path for existing part
->>>>> updateExistingEecacheEntry
 >>>>>> checkForEecacheEntryUpdate
 >>>>>>> setPartAndSerialNumberAttributes   <-- saves data from eecache
 >>>>>>> memcmp(eecache,ipz_record)
 >>>>>>> updateEecacheContentsFromBuffer
 >>>>>>> isEepromInSync(use_attrs=true)
 >>>>>>>> VPD::ensureEepromCacheIsInSync(use_attrs=true)

Change-Id: Ic9a026266e59ae14e9641fe5f274c55cb27d1c70
CQ: SW544254
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/129127
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>